TAL | BE | Web Services
TAL-264 | BE | Organization Web Service
Added constructor injection in the resource and the final Javadoc.

TAL-265 | BE | Organization Capability Web Service
Added constructor injection in the resource and the final Javadoc.

TAL-266 | BE | Skill Web Service
Added constructor injection in the resource and the final Javadoc.

TAL-268 | BE | Project Web Service
Added constructor injection in the resource and the final Javadoc.

TAL-269 | BE | Capability Web Service
Added constructor injection in the resource and the final Javadoc.