TAL-298 | BE | Token Authentication
Created the TokenAuthenticationFilter, its service interface and implementation. Modified the User object in the model to add the token attribute to each User. Modified the spring security configuration files to include the token authentication filter. Changed the POM files to add the javax servlet dependency to the root POM with another version because the other one was giving problems with the AbstractAuthenticationProcessingFilter. Changed the TechnicalResource service and dao to include necessary changes.