TAL-327 BE | Contact Us
Removed the 'subject' property of ContactEmail.java
Added the enum IssueType to keep track of the type of issue in a contactEmail, and added it as an attribute of ContactEmail.java
Made the necessary changes in the test of ContactEmail to account for the changes made to the class.
Added dao and service for the ContactEmail entity.
Added a web service that creates and persists a ContactEmail in ContactEmailResource.java, as well as the pattern to its endpoint in talent-web-security.spring.xml.