TAL-310 | BE | Organization Filter Authentication
Implemented the organzization authentication provider and organization filter, and replaced Spring's FORM_LOGIN_FILTER with it.
Changed TechnicalResource to have a unique constraint on the username and organization_id columns, instead of just username.
Changed the TechnicalResourceService to get users from both username and organization identifier.
Fixed HQL query in the Technical Resource DAO to correctly retrieve a user based on username and organization identifier.